Adds ignored title part of podcast urls
